Port More Recent PRs from upstream (#90)
* flake8 fixups

* Implement linting in CI

* Fix pandas FutureWarning, GH#5260

* dummy commit to test #5263

* Remove dummy file

* Port #5263

* Port #4925

* silence lambda warning

* Port #5042

* strict linting

* linting

* Port pieces of #4984

* Port pieces of #4985

* Port pieces of #4987

* comment

* Comments

* cleanup

* Port #5265

* Cleanup

* Port #4969

* Port #5234

* revert change that causes test failure

* Add flake8 dep

* Fixup trailing whitespace

+++ sm2/tsa/filters/tests/test_filters.py
@@ -120,7 +120,7 @@ def test_bking2d():
         [67.06026, -.1766731], [91.87247, .3898467], [124.4591, .8135461],
         [151.2402, .9644226], [163.0648, .6865934], [154.6432, .0115685]])
 
-    X = macrodata.load_pandas().data[['realinv', 'cpi']].values.astype(np.float)
+    X = macrodata.load_pandas().data[['realinv', 'cpi']].values.astype('f8')
     Y = bkfilter(X, 6, 32, 12)
     assert_almost_equal(Y, bking_results, 4)
